Started restructuring the namespace hierarchy. Things are all grouped under edu.zao.fire now. Below that will be 'editors', 'rcp', 'util', and 'views'. The rcp is just the stuff that is necessary to make the UI go (basically the auto-generated stuff from when I created the RCP project).
Editors will eventually be the pieces of the UI that you edit the Renaming Rules, and they will be the things that save and load renaming rules. The Views will be stuff like the file browser view, and potentially the metadata browser.

Util will just be little helpful things that could be used by anything in our program if it needed to be.

Looks like a lot of changes because the package names changed in pretty much every file.